logique métier de reception

+++ src/main/java/ma/baggar/bmsback/Entity/ReceptionDetail.java
@@ -14,16 +14,19 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Entity
 @Data @AllArgsConstructor @NoArgsConstructor
 public class ReceptionDetail {
 	@Id @GeneratedValue(strategy = GenerationType.IDENTITY)
 public Long id;
-	@ManyToOne @NotNull(message = "ajouter l article")
+	@ManyToOne
+	@NotNull(message = "ajouter l article")
 private Article article;
 	@NotNull(message = "il faut ajouter la quantité")
 private float quantite;
-private double prixTtc;
+private BigDecimal prixTtc;
 @ManyToOne(fetch = FetchType.LAZY, optional = false)
 @JoinColumn(name = "reception_id", nullable = false)
 @JsonIgnore
