+++ src/main/java/ma/baggar/bmsback/Service/ReceptionService.java
@@ -1,5 +1,6 @@
 package ma.baggar.bmsback.Service;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import ma.baggar.bmsback.Dto.ReceptionDto;
@@ -8,12 +9,10 @@ import ma.baggar.bmsback.Entity.ReceptionDetail;
 public interface ReceptionService {
 ReceptionDto saveReception(ReceptionDto receptionDto);
 List<ReceptionDto> getAllReception();
-double getPrixTtcOfArticle(double prixAchat,float quantite,double tvaValue);
-double calculePrixHt(List<Double> prixHt);
-double calculePrixTtc(List<Double> prixTtc,double fret,double remise);
-List<Double> getListOfPrixHtFromReceptionDetails(List<ReceptionDetail> receptionDetails );
-List<Double> getListOfPrixTtcFromReceptionDetails(List<ReceptionDetail> receptionDetails );
-
-
+BigDecimal getPrixTtcOfArticle(BigDecimal prixAchat, float quantite, double tvaValue);
+BigDecimal calculePrixHt(List<BigDecimal> prixHt);
+BigDecimal calculePrixTtc(List<BigDecimal> prixTtc,BigDecimal fret,BigDecimal remise);
+List<BigDecimal> getListOfPrixHtFromReceptionDetails(List<ReceptionDetail> receptionDetails );
+List<BigDecimal> getListOfPrixTtcFromReceptionDetails(List<ReceptionDetail> receptionDetails );
 
 }
